Moving fixtures to another file
